Not so pretty solution to functionality improvements.
Not so huge fun of so much exploded `loop/1` function.  It seems like
there is too much logic, even if it's only communication (adding ref's
and sending back messages).  So it's not so bad, but still it's hard
to read.  Some would say that reading Erlang is a skill, but it should
not be, it's to noisy, even if it's declarative.